Moves ElastiCache parameter expander and flattener to resource file
